Update MQTT_publisher.py
Having some problems with the publisher losing connection to the server so added some code to log errors to a file. Also refactored the connection function to limit the number of times connection is attempted. If connection fails an error is logged and the pico reset. The problem may be weak WiFi but I want to eliminate other possibilities.